Convert Auth into top level container App
